Revert "Move the DuckChatContextualDataStore interface to the API module to fix the lint issue"
This reverts commit c2a1a7b15be0483d3f4170ab1bf09e1f48c0bca2.
